test(language/compiler/check): update (some) check fixtures with new split annotated / checked snapshots

+++ language/compiler/src/check/tests/types/literals/import.rs
@@ -81,7 +81,7 @@ const copy = counter;
 === values.ds ===
 
 === annotated ===
-export let counter: float64 = 1 as float64;
+export let counter: float64 = 1;
 
 === checked ===
 export let counter = 1;
